pyproject.toml
[project]
name = "kada-commute"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-auth>=2.43.0",
    "gspread>=6.2.1",
]
